requirements.txt
aiogram==2.25.2
python-dotenv==1.1
mysql-connector-python==9.2
mysql==0.0.3
cryptography==44.0.2
aiohttp==3.8.6